jetstream1 branch changes

+++ getdata.py
@@ -45,15 +45,17 @@ def get_data(thefile):
 
 def download(url,newdir,newfile):
     filestring=newdir+newfile
-    urlstring="wget -O "+newdir+newfile+" "+url
-    print urlstring
+    if os.path.isfile(filestring):
+	print "file exists:",filestring
+    else:
+    	urlstring="wget -O "+newdir+newfile+" "+url
+   	print urlstring
     #Note: only for Python 3
     #urllib.request.urlretrieve(url,filestring)
     #Use this for Python 2
-    s=subprocess.Popen(urlstring,shell=True)
-    s.wait()
-    
-    print "Finished downloading from NCBI."
+    	s=subprocess.Popen(urlstring,shell=True)
+    	s.wait() 
+    	print "Finished downloading from NCBI."
 
 #3. Extract with fastq-dump (sratools)
     
@@ -90,37 +92,8 @@ def fastqc_report(fastq_file_list,newdir,fastqcdir,filename):
     	fastqc_string="fastqc -o "+fastqcdir+" "+file_string
     	print fastqc_string
     	print "fastqc reports generated for: "+str(fastq_file_list)
-    	#s=subprocess.Popen(fastqc_string,shell=True)
-    	#s.wait()
-
-#5. For pipeline testing only:
-#   create subset of 1,000,000 reads for each file
-
-def subset_reads(newdir,subsetdir):
-    listoffiles=os.listdir(newdir)
-    for i in listoffiles:
-	if i.endswith(".fastq"):
-		newfilename=subsetdir+i[:-6]+".subset1m.fastq"
-    		if os.path.isfile(newfilename):
-			print "File has already been subsetted:",newfilename
-    		else:	
-    			subset_string="head -4000000 "+newdir+i+" > "+newfilename
-    			print subset_string
-    			#s=subprocess.Popen(subset_string,shell=True,stdout=PIPE)
-    			#s.wait()
-			print "Finished subsetting."
-   
-#6. Create symbolic link from data files to working directory
-
-def sym_link(newdir):
-    listoffiles=os.listdir(newdir)
-    for i in listoffiles:
-    	if i.endswith(".subset100k.fastq"):
-    		symlink_string="ln -fs "+newdir+i+" /mnt/mmetsp/"+i
-		print symlink_string
-	    
-
-# this is the main function to execute
+    	s=subprocess.Popen(fastqc_string,shell=True)
+    	s.wait()
 
 def execute(basedir,url_data):
     for item in url_data.keys():
@@ -136,19 +109,21 @@ def execute(basedir,url_data):
             filename=basename(urlparse(url).path)
             print filename
             newdir=org_seq_dir+filename+"/"
+	    full_filename=newdir+filename
             clusterfunc.check_dir(newdir)
 	    fastqcdir=newdir+"fastqc/"
 	    clusterfunc.check_dir(fastqcdir)
             #check to see if filename exists in newdir
             if filename in os.listdir(newdir):
                print "sra exists:",filename
+	       if os.stat(full_filename).st_size == 0:
+                   print "SRA file is empty:",filename
+                   os.remove(full_filename)
             else:        
                print "file will be downloaded:",filename
 	       download(url,newdir,filename)
             sra_extract(newdir,filename)
-            #fastqc(newdir,fastqcdir,filename)
-            #subset_reads(newdir,subsetdir)
- 	    #fastqc(subsetdir,subsetfastqcdir,filename)
+            fastqc(newdir,fastqcdir,filename)
 
 def fastqc(newdir,fastqcdir,filename):
 	listoffiles=os.listdir(newdir)
@@ -159,22 +134,10 @@ def fastqc(newdir,fastqcdir,filename):
 			fastq_file_list.append(newdir+i)
 	fastqc_report(fastq_file_list,newdir,fastqcdir,filename)         
 
-# use this in case you make a mistake and need to delete a bunch of files:
-def delete_files(newdir):
-	listoffiles=os.listdir(newdir)
-        print listoffiles
-	for i in listoffiles:
-		if i.endswith(".subset100k.fastq"):
-			os.remove(newdir+i)
-			print "File removed:",newdir+i
-
-datafiles=["MMETSP_SRA_Run_Info_subset2.csv",
-	"MMETSP_SRA_Run_Info_subset_d.csv","MMETSP_SRA_Run_Info_subset_a.csv",
-	"MMETSP_SRA_Run_Info_subset_b.csv"]
-basedir="/mnt_redo/mmetsp/"
+datafile="MMETSP_SRA_Run_Info_subset_jetstream1.csv"
+basedir="/vol1/mmetsp/"
 clusterfunc.check_dir(basedir)
-for datafile in datafiles:
-	url_data=get_data(datafile)
-	print url_data
-	execute(basedir,url_data)
+url_data=get_data(datafile)
+print url_data
+execute(basedir,url_data)
